Finished reverse.html
POSTs a JSON dictionary with the key token and my token value to http://challenge.code2040.org/api/reverse
This endpoint returns a string to be reversed. We that string, then sends it back to http://challenge.code2040.org/api/reverse/validate, using a POST request containing the reversed string in a JSON dictionary.