Finished prefix.html
A POST request is sent to http://challenge.code2040.org/api/prefix and the server responds with a dictionary containing two keys and two values. The first value, prefix, is a string. The second value, array, is an array of strings. A result array is then created and all strings in the input array not beginning with the input prefix are added to that result array. Then the array is sent back to http://challenge.code2040.org/api/prefix/validate through another POST request.